+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Added `hitCondition` support for SourceBreakpoint and FunctionBreakpoint.
 - Added `isOptimized` support in Module.
 - Added `isUserCode` support in Module.
+- Added `symbolFilePath` support in Module.
 
 #### Changed
 - Replaced VSCode to DAP (variables, class names, tests, etc).